backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.3.0
requests==2.31.0
pdfplumber==0.10.3
python-dotenv==1.0.0
arxiv==2.1.0
google-generativeai
groq
openai
faiss-cpu
sentence-transformers
numpy
scipy
transformers
torch
pyyaml
aiofiles
mysql-connector-python==8.2.0
email-validator
